more improvements

+++ ping_targets/dst_ping.sh
@@ -9,13 +9,15 @@ TARGETS="AntonHomeSW HomeStolbSW GaponStolbRT MaslukStolbRT TulnovaHomeRT"
 function push_slack() {
         if [[ -n "$results" ]] && [[ ! -f tmp.txt ]] ; then
                 STATUS=warning
-                echo $STATUS > tmp.txt
+                echo $result > tmp.txt
 		echo "$results" |grep -E  "failure|denied" > /dev/null && STATUS=danger
                 echo "$results" |/usr/local/bin/slacktee.sh -u "$HOST" -t "$APP" -a $STATUS -e "cron date" "$(date)"
-        else    
-                echo "Connection restored $results" | \
-                /usr/local/bin/slacktee.sh -u "$HOST" -t "$APP" -a good -e "cron date" "$(date)"
-                rm tmp.txt
+        else 
+                if [ -f tmp.txt ] ; then
+                        echo "Connection restored $results" | \
+                        /usr/local/bin/slacktee.sh -u "$HOST" -t "$APP" -a good -e "cron date" "$(date)"
+                        rm tmp.txt
+                fi
         fi
 }
